chore(canvas) Move drag-to-escape-hatch code into escape hatch strategy file
