+++ editor/src/components/canvas/controls/select-mode/canvas-strategy-picker.tsx
@@ -14,9 +14,7 @@ export const CanvasStrategyPicker = React.memo(() => {
   const { allApplicableStrategies } = useEditorState(
     Substores.restOfStore,
     (store) => ({
-      allApplicableStrategies: store.strategyState.sortedApplicableStrategies?.filter(
-        (strategy) => strategy.strategy.fitness > 0,
-      ),
+      allApplicableStrategies: store.strategyState.sortedApplicableStrategies,
     }),
     'CanvasStrategyPicker strategyState.currentStrategy',
   )
